module_03_rag/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.0",
    "ddgs>=9.10.0",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.10",
    "langchain-classic>=1.0.1",
    "langchain-core>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.9",
    "langchain-qdrant>=1.1.0",
    "pandas>=3.0.0",
    "pymupdf>=1.27.1",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.2.2",
    "pypdf>=6.7.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
]
